backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-multipart==0.0.19
python-dotenv==1.0.1
scikit-learn==1.6.1
numpy==2.2.1
pydantic[email]==2.10.4
